Update composer.json to use latest Robo helpers
In an effort to upgrade Laravel in opening hours, we stumbled upon a
dependency on `digipolis-robo-helpers` (4.x) which in turn depended on
`roave/better-reflection`, which depends on PHP 8.0 - 8.2, not 8.3,
making it a drag to upgrade.

We are now using the 5.0.. version of Robo helpers and at the same time
we locked down the PHPUnit version, in line with
`robo-digipolis-drupal`.